+++ mbgdml/models/__init__.py
@@ -1,0 +1,5 @@
+
+from .base import model
+from .gap_model import gapModel
+from .gdml_model import gdmlModel
+from .schnet_model import schnetModel